+++ src/application/handlers/mutation/Note/NoteDeleteHandler.ts
@@ -5,7 +5,14 @@ import { DeleteResult, MutationHandlerFunc } from "../../../types/Handlers";
 export const NoteDeleteHandler: MutationHandlerFunc<
   Note,
   DeleteResult
-> = async (payload: Pick<Note, "id">, prisma: PrismaClient, user: User) => {
+> = async (
+  payload: Pick<Note, "id">,
+  prisma: PrismaClient,
+  user: User,
+  schema
+) => {
+  await schema.validateAsync(payload);
+
   const note = await prisma.note.findUnique({
     where: {
       id: payload.id,
@@ -15,6 +22,7 @@ export const NoteDeleteHandler: MutationHandlerFunc<
     },
   });
 
+  if (!note) throw new GraphQLError("Not Found");
   if (user.username !== note.user.username) throw new GraphQLError("Forbidden");
 
   const deleted = await prisma.note.delete({
